mesa: add reverse lookup map for non-namespace paths
this map is useful to track what namepace path different listeners
are using when, for example, using the same internal path (i.e. peeking
for the same path using the %publ namespace with -meen, but also
with the per-peer %chum namespace used in -chum).

Here we track also what types of interest gifts different listeners expect.
Originally this was either %tune (for |fine) or %sage (for |mesa) but we also
add a %rate gift so subscribers can be notified of the progress of the +peek.

(WIP: %yawn/%wham paths to cancel peeks)