tests: track guest vs real eyre session status
In the eyre tests, we already auto-track the session token from the
cookies. Recent changes however, made it so that the cookie responses
might have different expiry times based on what _kind_ of session it is.

So, we add a flag to test harness state that says whether we're using a
guest session right now or not. We could store the full $identity:eyre,
but don't strictly need to right now, so we don't bother.

Based on the flag, the automated checks for the cookie fill in the
appropriate expiry time.