eyre: accept session token from request header
Cookies are a staple of the "HTTP web stack". As such, middleware
commonly handles receiving, storing and sending cookies for the
developer. Browsers do this fine, and eyre is tuned to their cookie
handling behavior.

Some middleware may handle cookies poorly, and not give the developer
sufficient control to mitigate the issues it causes. The tough love
answer is that the developer should use better middleware. While that
advice comes from a place of love, it also comes from a very privileged
position.

Here we add support for passing the eyre session token through the
`Authorization` header instead. We use the `Bearer` scheme because it
lets us put the existing session tokens right there: `@uv` notation
already limits itself to the base64 character set, we can simply take
the bearer token string as-is.

If an `Authorization` header is provided, any cookies in the request are
ignored. If no `Authorization` header is provided, the request is
handled as normal.

Additionally, responses to successful login requests will now contain
the session token in their response bodies. The optional redirect
behavior remains unchanged.

We include a test for this new authentication (authorization?) method.