eyre: initial desk segregation prototype
Userspace permissions at the gall level is somewhat fake, because at the
eyre level none of those permissions get enforced. Making eyre do so
requires browse/http clients to specify which desk they
represent/operate as, so that the appropriate permissions can be
enforced.

Making such a setup actually secure (preventing desks other other
client-side code from pretending to be each other) is a bit of a rabbit
hole. Here, we lean into an approach originally proposed by
@philipcmonk, which can be found here:

https://gist.github.com/fang-/18ce946a2bb33bada3f10bdb3546bb55

This commit introduces changes to eyre that make it look up the
corresponding desk based on the request path, then serve a page
containing an iframe that points to the subdomain for that desk (at the
same path). This proves out the basic setup, but doesn't yet implement
any automated authentication flows or other UX necessities, and doesn't
even begin to solve for subdomain SSL requirements.

We will continue iterating on this. The code here is experimental and
liable to heavy changes. Better documentation of the approach taken,
cases considered and the designed behavior will follow once the
implementation settles.

For development log and discussion notes, see the Eyre Security Working
Group meeting notes.

https://gist.github.com/fang-/b2b405610039bd04a69464125a3706ec