eyre: various desk segregation todos
Handles a handful of TODOxx comments introduced in the parent commit.

We now effectively assert the presence of a Host header.

We reify %gen bindings as "belonging" to the desk that contains them,
similar to how %app bindings "belong" to the agent they point to

Most importantly, we settle on the idea that EAuth and guest sessions
should behave no different from local sessions. In other words,
regardless of your login method, you are expected to pass scope-relevant
cookies, and are expected to obtain those through the (yet
unimplemented) flow dedicated to that. This implies that guest sessions
should only get auto-minted for requests to the "root" scope (at the
top-level domain), and not for requests elsewhere.

We continue to hack around eyre's poorly factored request handling.