eyre: refactor request authentication handling
We move the +session-for-request arm to be inline inside of the +request
handling function. This makes that already-long chain of logic even
longer, maybe we split it back out later, but the important change here
is that we explicate the result from "seeking and maybe creating" an
authentication session on the incoming request.

We lightly update the logic to handle the newly-shaped response
correctly but don't otherwise change the behavior. (Aside from adding an
explicit %negotiate case, which the follow-up commit will handle.)