eyre: negotiate auth between subdomains and frame
When the user opens an app for the first time, they will generally load
it in from its subdomain through the outer frame that they interact with
their urbit from.

Their browser will not have any authentication tokens or cookies for the
subdomain yet, but they're authentiated into the top-level domain (in
the outer frame).

Here, we make handling of the subdomain request serve an "authentication
negotiation script" when the incoming request is fully unauthenticated.
This script sends a message to the outer frame (if present) asking for a
fresh cookie. The outer frame fetches it (or will eventually, work in
progress) and injects it into the inner frame, so that it may add it to
browser state.

Plenty of TODOs inside, but the overall setup here is implemented and
has been observed to work.