ames: migrate all nacked sequence numbers
migrating to mesa removes state by deleting the naxplanation flows. in order to keep as much state as possible, we try to migrate every sequence number for which we can infer that it was indeed nacked.

nax.message-sink-state existe but only keeps messages there if the nacked sequence number is gth > 1.

here we infer that by finding the reference flow =(%0 (mod bone 4)) of a naxplanation flow =(%03 (mod bone 4)) and assert that there are as many messages (plus one, since the message pump has always one message more than the message sink) recevied as we have sent naxplanations.

(this might help in cases where we are peeking for naxplanation and the peek could block if the other side has not migrated the nacked sequence numbers)