eyre: add and use scoped token minting endpoint
At the /~/login endpoint, making POST requests that contain a
"xxtokenflow" and "scope" argument will now mint a new scoped session
token and return that in the response body.

We make the outer frame script fetch this when the inner frame asks, and
the inner frame writes the resulting token into a cookie.

With this, basic happy-path automagic subdomain authentication seems to
work!