iris: handle requests when we don't yet have a duct
Recently the %groups desk started doing http requests to posthog during
+on-init. This works fine when upgrading %groups but fails miserably when
booting from a pill, crashing in arvo and bringing the ship down with
%give-no-duct.

To ameliorate the issue we modify %iris to not give anything if the duct is
still empty. In practice this results in the caller getting a %cancel every time
if the http request happens early during the bootstrap process, since all
outstanding http requests get canceled when the %born event comes in later from
vere. This is obviously not good but it's better than a broken pill and a
canceled http request will not matter in this specific instance for the posthog request.