hoon: remove doccords types
Since we no longer parse for them, we can rip them out entirely. This
gives us approximately 15% perf improvement on compilation.

The most invasive part of this change is taking the $what off of $tomes,
which affects a lot of code, but mostly in very trivial ways.

Note that we retain a nesting sample type on +loot, because that's
jetted. This way we don't have to coordinate a jet update. We eat a
+run:by to make this work, but those maps shouldn't ever be huge in
practice.