eyre: drop iframes, switch to redirect approach
The iframe approach doesn't work due to "browser" "security" "behavior".
We take a step back and explore what it would be like to have the
subdomains visible to the user.

When visiting a subdomain, if no authentication is provided, we redirect
back to the top-level domain at /~/xxauth, which receives the target
subdomain+path in the url, and mints a temporary token for us that's
associated with whatever identity we have on the top-level domain.

That then redirects us back to the subdomain, at /~/xxauth, with that
temporary token and target path in the url. Eyre validates the temporary
token, grants us a scoped cookie, and serves the final redirect.

The factoring as of this commit isn't great. We also don't have real
tokens or validation yet, and always mint a %local cookie. But: the
redirects work smoothly, and don't even show up in browser history.

Further work to be done here, obviously.