hoon: allow comments at end-of-file
Prior to removing doccords, presumably due to doccords parser behavior,
we allowed comments at the direct end of file, without a trailing
newline. 60af617 broke this behavior.

It's probably sane to allow comments without a trailing newline in
general, so here we update the comment parser +vul to either parse a
newline or, failing that, check for end of the string.