hoon: UIP-135 style date parsing and rendering
Previously, the `@da` literal syntax never allowed leading zeroes in its
month and day segments. In its hour, minute and second segments however,
it allows leading zeroes during parsing, and actively includes them
during rendering to ensure consistent double-digit representations.

Here, we extend the leading zero behavior of the time segments into the
`@da` format's month and day segments for both parsing and rendering,
accepting and displaying `~2025.1.1` as `~2025.01.01`.

This is an implementation of the specification in UIP-135, see that for
additional context.