eyre: send cookie expiry when %negotiate
Sometimes when we end up with %negotiate auth-state, this means a cookie
_was_ provided but that it's invalid. In those cases, we should take
care to include the same behavior that those cases have on root domain,
which is to send an expiry notice of the cookie to the client.

Here, roll an optional "old session" into the %negotiate auth-state and
send the appropriate set-cookie header when serving its response.

This should also help get you out of redirect loops in cases of multiple
cookies being sent by the browser, where the "correct" one is further
down the stack. For multi-cookie support in general, we probably want
more and/or a different track of work.