gall: kernel resource tracking prototype
We modify gall to track kernel resources used by agents. To do this, we
narrow and redefine the kernel api that agents have access to. This way,
gall always has enough information to unique address a resource in the
way that its corresponding vane expects.

We redefine some of the lull types and stub out state migrations
entirely to ease development for the time being. Whenever this work
finalized, we'll need to update all of the default agents, write
migrations, etc.

Includes basic tests for resource tracking, that assume this version of
gall lives at /gall.hoon in the base desk.