eyre: only serve login page on root domain
And redirect to subdomains post-login if desired.

We add an additional URL parameter for the login page, `desk`, which
specifies the subdomain to redirect to iff a `redirect` param is also
given.

Note that we actually redirect to xxauth in that case, because login
success means we'll want a new scoped cookie for the redirect target
anyway.