eyre: construct host string with port consistently
Add a +host-string helper for constructing a host from a [turf port]
tuple. This makes it obvious that we weren't always including the port,
which would break redirects in some cases.

Some inconsistency in host string handling remains (mainly stemming from
reusing the `@t` host instead of working with the [turf port]), but
aligning all of that is out of scope for the time being.