gall: emit %raw-fact in %unto ova
When an %unto gift is emitted by gall, no care is taken to detect
whether the gift will turn into an effect. This makes it impossible
to receive facts externally, such as when using aqua, because
they carry a vase. Type for a vase with a type containing a %hold
usually can not be verified efficiently, and thus one can't receive a fact at all, as molding such a noun will require a long time to compute.

To fix this, we detect in gall whether a fact is about to be transferred to the outside. If so, we convert the %fact into a %raw-fact, which does not carry a vase.