gall: integrate remote scry api
Remote scry resource tracking already existed in gall prior to our
changes. All we needed to do was move the API for it around so it fit
into our new model, with the $task-user-v1 and such. However, the
factoring was confusing/annoying, so we ended up doing a bunch of
fiddling along the way.

- We clarify %huck as being used only for kicking subscriptions.

- We introduce %jump into $neet as an "out" for gall when describing
  kernel tasks on the agent's behalf. You only need that when doing
  card->move transformations prior to calling +ap-from-internal, which
  only happens in +ap-handle-result (for remote scry card transforms).

- We fix the ordering of scry request tracking wrt +on-arvo callbacks.

- We manage to delete +ap-handle-ken by merging its logic into
  +ap-handle-sky, with which it was already partially redundant.

We do not yet fix a bug we've identified in +ap-reinstall, which may
incorrectly do a normal scry for encrypted data.

No tests yet, but would be very good to have for this.