ames, gall: try to %flub in-progress $pleas
Prior to the introduction of the %flub $gift, %gall would enqueue all incoming %pleas to non-running or suspended agents, and defer a %done to %ames until the agent is %live. For any of these $pleas, %ames would remember that it had heard this message, and when a retry of the last fragment would arrive, it would drop it on the floor, knowing that %gall has received this and enqueued it already, and we are just waiting for the %ack.

This skips the current %flub code path, so we are never halting these flows.

In this PR we pass a special %plea (/gp path) that will check if the agent is running, and if not, give a %flub to %ames.