hoon: make parsers operate on cords
Changes the parsers' "continuation" from a tape representinig the
remainder of the parse input, to the original parse input paired with a
pointer into it.

Includes a rewritten +inde (and deduplicates +iny by pointing to it
explicitly), because the original implementation was difficult to make
work in this new context, and was somewhat questionable anyway.
(Degenerate cases would reparse the entire rest of the input...)

Note that for converting inputs from tapes, we must be cautious around
two things:
- +crip would drop null characters, so we must +rep instead.
- Trailing null characters cannot be represented by the cord itself so
  must instead be represented by the .m in the parser input. Therefore,
  when converting from tape, measure the tape length first, instead of
  measuring the resulting atom.

Despite changes, /lib/der gets broken here. To be fixed at a later date
if necessary.

Must come paired with updated jets.

From initial measurements, performance with these changes remains mostly
unchanged, so the juice might not be worth the squeeze here.